backend/requirements.txt
fastapi
uvicorn
logging
pymongo
pika
pydantic
python-dotenv
